+++ src/EventSubscriber/TranslatableEventSubscriber.php
@@ -6,9 +6,10 @@ namespace Knp\DoctrineBehaviors\EventSubscriber;
 
 use Doctrine\Bundle\DoctrineBundle\Attribute\AsDoctrineListener;
 use Doctrine\ORM\Event\LoadClassMetadataEventArgs;
+use Doctrine\ORM\Event\PostLoadEventArgs;
+use Doctrine\ORM\Event\PrePersistEventArgs;
 use Doctrine\ORM\Events;
 use Doctrine\ORM\Mapping\ClassMetadataInfo;
-use Doctrine\Persistence\Event\LifecycleEventArgs;
 use Doctrine\Persistence\ObjectManager;
 use Knp\DoctrineBehaviors\Contract\Entity\TranslatableInterface;
 use Knp\DoctrineBehaviors\Contract\Entity\TranslationInterface;
@@ -62,14 +63,14 @@ final class TranslatableEventSubscriber
         }
     }
 
-    public function postLoad(LifecycleEventArgs $lifecycleEventArgs): void
+    public function postLoad(PostLoadEventArgs $postLoadEventArgs): void
     {
-        $this->setLocales($lifecycleEventArgs);
+        $this->setLocales($postLoadEventArgs);
     }
 
-    public function prePersist(LifecycleEventArgs $lifecycleEventArgs): void
+    public function prePersist(PrePersistEventArgs $prePersistEventArgs): void
     {
-        $this->setLocales($lifecycleEventArgs);
+        $this->setLocales($prePersistEventArgs);
     }
 
     /**
@@ -154,9 +155,9 @@ final class TranslatableEventSubscriber
         }
     }
 
-    private function setLocales(LifecycleEventArgs $lifecycleEventArgs): void
+    private function setLocales(PostLoadEventArgs|PrePersistEventArgs $eventArgs): void
     {
-        $object = $lifecycleEventArgs->getObject();
+        $object = $eventArgs->getObject();
         if (! $object instanceof TranslatableInterface) {
             return;
         }