+++ CHANGELOG.md
@@ -15,6 +15,7 @@ PRs and issues are linked, so you can find more about it. Thanks to [ChangelogLi
 
 - Update for newer versions of symfony, php, doctrine and etc. after fork original KnpLabs/DoctrineBehaviors [@schvoy]
 - Remove deprecations [@schvoy]
+- Replace LifecycleEventArgs with specific EventArgs class based on the event. [@schvoy]
 
 ## [v2.0.1] - 2020-01-08